Add database facility
This List project is now able to store the items entered by user to database with the help of MongoDB and mongoose. In addition to adding the database support, it has now the feature of deleting items from the database, you just need to click on the checkbox of the item that you want to delete!